+++ src/Edge.h
@@ -6,15 +6,15 @@ class Edge
     private:
         unsigned int fromID, toID;
         float weight;
-        short edgeID;
+        unsigned int edgeID;
     
     public:
 
         // Constructor
-        Edge(unsigned int fromID, unsigned int toID, float weight, short edgeID);
+        Edge(unsigned int fromID, unsigned int toID, float weight, unsigned int edgeID);
 
         // Accessors
-        short getEdgeID() const;
+        unsigned int getEdgeID() const;
         unsigned int getFrom() const;
         unsigned int getTo() const;
         float getWeight() const;
